refactor(#640): rename sheet_dsl.py → sheet.py — the Sheet facade owns the natural name
With the layout engine renamed to compose.py, the module holding the
user-facing Sheet facade takes the sheet.py name. The old name was also
a quiet contradiction: ADR 0001's headline decision was *against* an
editable DSL, and ADR 0011 deliberately calls this surface a fluent
facade — yet the module said dsl.

- draftwright.sheet_dsl stays as a frozen deprecation shim (module
  __getattr__ forwarding + DeprecationWarning) for one release, with a
  regression test.
- Public API unchanged: draftwright.Sheet re-exports as before; emitted
  --script files already import from the package root.
- linting/coverage.py comment cross-references and CLAUDE.md updated.
