+++ src/draftwright/__init__.py
@@ -30,7 +30,7 @@ _LAZY = {
     "make_drawing": "draftwright.builder",
     "Drawing": "draftwright.drawing",
     "FeatureInfo": "draftwright.drawing",
-    "Sheet": "draftwright.sheet_dsl",
+    "Sheet": "draftwright.sheet",
     "fix_svg_page_size": "draftwright.export",
     "lint_feature_coverage": "draftwright.linting",
     "PmiRecord": "draftwright.pmi",
@@ -80,7 +80,7 @@ if TYPE_CHECKING:  # static analysers / IDEs — no runtime import, no kernel co
     from draftwright.linting import lint_feature_coverage
     from draftwright.pmi import PmiRecord, extract_pmi
     from draftwright.recognition import recognise_face_levels
-    from draftwright.sheet_dsl import Sheet
+    from draftwright.sheet import Sheet
 
 
 def __dir__():
